requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
av==14.3.0
babel==2.17.0
beautifulsoup4==4.12.3
bleach==6.2.0
braceexpand==0.1.7
Brotli==1.0.9
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.5.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
files-to-prompt==0.6
fonttools==4.55.8
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2024.12.0
fvcore==0.1.5.post20221221
gdown==5.2.0
gmpy2==2.1.2
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
humanize==4.12.2
idna==3.7
imageio==2.37.0
importlib_resources==6.5.2
iopath==0.1.10
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lazy_loader==0.4
libauc==1.4.0
littleutils==0.2.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
monai==1.4.0
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.3.2
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
ogb==1.3.6
opencv-python==4.11.0.86
outdated==0.2.2
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pip==24.2
platformdirs==4.3.6
portalocker==3.1.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==6.30.0
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytorchvideo==0.1.5
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.22.3
safetensors==0.5.2
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.1.0
shellingham==1.5.4
SimpleITK==2.4.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.1.10
timm==1.0.14
tinycss2==1.4.0
tomli==2.2.1
torch==2.5.1
torch-geometric==2.6.1
torchaudio==2.5.1
torchio==0.20.8
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typer==0.15.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.11.1
webdataset==0.2.111
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.17.2
XlsxWriter==3.2.2
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.0
